Don't replace whitelist-ips.conf and whitelist-domains.conf if they already exist

+++ ls-bad-bot-blocker/sbin/deploy.sh
@@ -11,8 +11,12 @@ wget -O /etc/nginx/bots.d/blockbots.conf                    "https://raw.githubu
 wget -O /etc/nginx/bots.d/ddos.conf                         "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/ddos.conf"
 
 #STEP 3:
-wget -O /etc/nginx/bots.d/whitelist-ips.conf                "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/whitelist-ips.conf"
-wget -O /etc/nginx/bots.d/whitelist-domains.conf            "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/whitelist-domains.conf"
+if [ ! -f /etc/nginx/bots.d/whitelist-ips.conf ]; then
+    wget -O /etc/nginx/bots.d/whitelist-ips.conf                "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/whitelist-ips.conf"
+fi
+if [ ! -f /etc/nginx/bots.d/whitelist-domains.conf ]; then
+    wget -O /etc/nginx/bots.d/whitelist-domains.conf            "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/whitelist-domains.conf"
+fi
 ## to do
 # populate 
 # /etc/nginx/bots.d/whitelist-ips.conf
@@ -20,22 +24,22 @@ wget -O /etc/nginx/bots.d/whitelist-domains.conf            "https://raw.githubu
 
 #STEP 4: BLACKLIST USING YOUR OWN CUSTOM USER-AGENT BLACKLIST
 if [ ! -f /etc/nginx/bots.d/blacklist-user-agents.conf ]; then
-wget -O /etc/nginx/bots.d/blacklist-user-agents.conf        "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/blacklist-user-agents.conf"
+    wget -O /etc/nginx/bots.d/blacklist-user-agents.conf        "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/blacklist-user-agents.conf"
 fi
 
 #STEP 5: BLACKLIST USING YOUR OWN CUSTOM BAD REFERRERS
 if [ ! -f /etc/nginx/bots.d/custom-bad-referrers.conf ]; then
-wget -O /etc/nginx/bots.d/custom-bad-referrers.conf         "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/custom-bad-referrers.conf"
+    wget -O /etc/nginx/bots.d/custom-bad-referrers.conf         "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/custom-bad-referrers.conf"
 fi
 
 #STEP 6: BLACKLIST IPS AND IP RANGES USING YOUR OWN CUSTOM LIST
 if [ ! -f /etc/nginx/bots.d/blacklist-ips.conf ]; then
-wget -O /etc/nginx/bots.d/blacklist-ips.conf                "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/blacklist-ips.conf"
+    wget -O /etc/nginx/bots.d/blacklist-ips.conf                "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/blacklist-ips.conf"
 fi
 
 #STEP 7: DOWNLOAD CUSTOM BAD REFERRER WORDS INCLUDE FILE FOR CUSTOMIZED SCANNING OF BAD WORDS
 if [ ! -f /etc/nginx/bots.d/bad-referrer-words.conf ]; then
-wget -O /etc/nginx/bots.d/bad-referrer-words.conf           "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/bad-referrer-words.conf"
+    wget -O /etc/nginx/bots.d/bad-referrer-words.conf           "https://raw.githubusercontent.com/layershift/nginx-ultimate-bad-bot-blocker/master/bots.d/bad-referrer-words.conf"
 fi
 
 #STEP 8: INCLUDE IMPORTANT SETTINGS IN NGINX.CONF
